run VxDesign's remaining scripts from compiled output
The twelve operational scripts (`create-organization`, `insert-dev-data`, `send-welcome-email`, and so on) ran as `node -r esbuild-runner/register ./scripts/x.ts` and all failed with `Unknown file extension ".ts"`. `-r` is a CommonJS preload, so it cannot help the ESM loader, and node routes a `.ts` entry point to that loader once the package is `"type": "module"`. No CI job runs them, so nothing went red.

They now run compiled: their modules move from `scripts/` to `src/scripts/` — which is what gets built, and where this package's other two CLI modules already live — and each package script becomes `pnpm build:self && node ./build/scripts/x.js`. Building first keeps them one command to run, as `scripts/README.md` documents them, and means they cannot silently execute a stale build against a real database. `tsc --build` is incremental, so the added step is a no-op once warm.

Only `from '../src/...'` specifiers changed inside the moved files. `scripts/` keeps its shell scripts, and `scripts/README.md` needs no change because it refers to scripts by `pnpm` name rather than by path.

With this, every entry point in this package works under ESM: the server, the worker, and all fourteen CLIs.

Verified by running all fourteen. Eleven print usage or do real work against a local Postgres — `list-organizations` returns rows, `insert-dev-data` reports the organization already present, `cancel-background-task` reports no task in progress — and `mock-circleci-server` binds and serves. `pnpm list-organizations` was also run end to end to exercise the rewritten package script. `lint` and `tsc` clean, suite 192/192 across 23 files with coverage unchanged, and `script/validate-monorepo` passes.
