+++ apps/design/backend/package.json
@@ -20,14 +20,14 @@
     "build:dev": "pnpm --filter $npm_package_name... build:self",
     "build:prod": "pnpm --filter-prod $npm_package_name... build:self",
     "build:self": "tsc --build tsconfig.build.json",
-    "cancel-background-task": "node -r esbuild-runner/register ./scripts/cancel_background_task.ts",
+    "cancel-background-task": "pnpm build:self && node ./build/scripts/cancel_background_task.js",
     "clean": "pnpm --filter $npm_package_name... clean:self",
     "clean:self": "rm -rf build && tsc --build --clean tsconfig.build.json",
-    "create-jurisdiction": "node -r esbuild-runner/register ./scripts/create_jurisdiction.ts",
-    "create-jurisdiction-user": "node -r esbuild-runner/register ./scripts/create_jurisdiction_user.ts",
-    "create-organization": "node -r esbuild-runner/register ./scripts/create_organization.ts",
-    "create-organization-user": "node -r esbuild-runner/register ./scripts/create_organization_user.ts",
-    "create-support-user": "node -r esbuild-runner/register ./scripts/create_support_user.ts",
+    "create-jurisdiction": "pnpm build:self && node ./build/scripts/create_jurisdiction.js",
+    "create-jurisdiction-user": "pnpm build:self && node ./build/scripts/create_jurisdiction_user.js",
+    "create-organization": "pnpm build:self && node ./build/scripts/create_organization.js",
+    "create-organization-user": "pnpm build:self && node ./build/scripts/create_organization_user.js",
+    "create-support-user": "pnpm build:self && node ./build/scripts/create_support_user.js",
     "db:migrations:ci-validate": "./scripts/ci-validate-migrations.sh",
     "db:migrations:create": "node-pg-migrate create",
     "db:migrations:run": "node-pg-migrate up",
@@ -39,14 +39,14 @@
     "db:schema": "pg_dump --schema-only -d",
     "db:schema-dev": "pg_dump --schema-only 'postgres://design:design@localhost:5432/design'",
     "format": "prettier '**/*.+(js|jsx|ts|tsx|css|graphql|json|less|md|mdx|sass|scss|yaml|yml)' --write",
-    "insert-dev-data": "node -r esbuild-runner/register ./scripts/insert_dev_data.ts",
+    "insert-dev-data": "pnpm build:self && node ./build/scripts/insert_dev_data.js",
     "lint": "pnpm type-check && eslint --cache .",
     "lint:fix": "pnpm type-check && eslint --cache . --fix",
-    "list-jurisdictions": "node -r esbuild-runner/register ./scripts/list_jurisdictions.ts",
-    "list-organizations": "node -r esbuild-runner/register ./scripts/list_organizations.ts",
-    "list-user-jurisdictions": "node -r esbuild-runner/register ./scripts/list_user_jurisdictions.ts",
-    "mock-circleci-server": "node -r esbuild-runner/register ./scripts/mock_circleci_server.ts",
-    "send-welcome-email": "node -r esbuild-runner/register ./scripts/send_welcome_email.ts",
+    "list-jurisdictions": "pnpm build:self && node ./build/scripts/list_jurisdictions.js",
+    "list-organizations": "pnpm build:self && node ./build/scripts/list_organizations.js",
+    "list-user-jurisdictions": "pnpm build:self && node ./build/scripts/list_user_jurisdictions.js",
+    "mock-circleci-server": "pnpm build:self && node ./build/scripts/mock_circleci_server.js",
+    "send-welcome-email": "pnpm build:self && node ./build/scripts/send_welcome_email.js",
     "start": "./scripts/start.sh",
     "test": "is-ci test:ci test:watch",
     "test:ci": "pnpm db:migrations:ci-validate && TZ=America/Anchorage ./scripts/test_ci.sh",
