+++ apps/design/backend/src/scripts/cancel_background_task.ts
@@ -1,6 +1,6 @@
 import { loadEnvVarsFromDotenvFiles } from '@votingworks/backend';
 import { BaseLogger, LogSource } from '@votingworks/logging';
-import { Db } from '../src/db/db.js';
+import { Db } from '../db/db.js';
 
 async function main(): Promise<void> {
   loadEnvVarsFromDotenvFiles();
