+++ apps/design/backend/src/scripts/create_jurisdiction.ts
@@ -4,10 +4,10 @@ import { resolve } from 'node:path';
 import { BaseLogger, LogSource } from '@votingworks/logging';
 import { assertDefined } from '@votingworks/basics';
 import { safeParse, SoftwareVersionSchema } from '@votingworks/types';
-import { createWorkspace } from '../src/workspace.js';
-import { WORKSPACE } from '../src/globals.js';
-import { generateId } from '../src/utils.js';
-import { Jurisdiction, StateCodeSchema } from '../src/types.js';
+import { createWorkspace } from '../workspace.js';
+import { WORKSPACE } from '../globals.js';
+import { generateId } from '../utils.js';
+import { Jurisdiction, StateCodeSchema } from '../types.js';
 
 const USAGE = `Usage: pnpm create-jurisdiction --organizationId=<organizationId> --stateCode=<stateCode> --softwareVersion=<softwareVersion> "<name>"`;
