+++ apps/design/backend/src/scripts/create_jurisdiction_user.ts
@@ -3,9 +3,9 @@ import util from 'node:util';
 import { assertDefined } from '@votingworks/basics';
 import { resolve } from 'node:path';
 import { BaseLogger, LogSource } from '@votingworks/logging';
-import { Auth0Client } from '../src/auth0_client.js';
-import { createWorkspace } from '../src/workspace.js';
-import { WORKSPACE } from '../src/globals.js';
+import { Auth0Client } from '../auth0_client.js';
+import { createWorkspace } from '../workspace.js';
+import { WORKSPACE } from '../globals.js';
 
 const USAGE = `Usage: pnpm create-jurisdiction-user --jurisdictionId=<string> <email address>`;
