+++ apps/design/backend/src/scripts/create_organization.ts
@@ -3,10 +3,10 @@ import util from 'node:util';
 import { resolve } from 'node:path';
 import { BaseLogger, LogSource } from '@votingworks/logging';
 import { assertDefined } from '@votingworks/basics';
-import { createWorkspace } from '../src/workspace.js';
-import { WORKSPACE } from '../src/globals.js';
-import { generateId } from '../src/utils.js';
-import { Organization } from '../src/types.js';
+import { createWorkspace } from '../workspace.js';
+import { WORKSPACE } from '../globals.js';
+import { generateId } from '../utils.js';
+import { Organization } from '../types.js';
 
 const USAGE = `Usage: pnpm create-organization "<name>"`;
