+++ apps/design/backend/src/scripts/insert_dev_data.ts
@@ -4,10 +4,10 @@ import { assertDefined } from '@votingworks/basics';
 import { BaseLogger, LogSource } from '@votingworks/logging';
 
 import { LATEST_SOFTWARE_VERSION } from '@votingworks/types';
-import { DEV_USER_ID } from '../src/auth0_client.js';
-import { NODE_ENV, votingWorksOrganizationId, WORKSPACE } from '../src/globals.js';
-import { Organization, StateCodes } from '../src/types.js';
-import { createWorkspace } from '../src/workspace.js';
+import { DEV_USER_ID } from '../auth0_client.js';
+import { NODE_ENV, votingWorksOrganizationId, WORKSPACE } from '../globals.js';
+import { Organization, StateCodes } from '../types.js';
+import { createWorkspace } from '../workspace.js';
 
 /**
  * Inserts the dev user needed when using AUTH_ENABLED=FALSE to bypass Auth0 in development as well
