+++ apps/design/backend/src/scripts/send_welcome_email.ts
@@ -1,6 +1,6 @@
 import { loadEnvVarsFromDotenvFiles } from '@votingworks/backend';
 import util from 'node:util';
-import { Auth0Client } from '../src/auth0_client.js';
+import { Auth0Client } from '../auth0_client.js';
 
 const USAGE = `Usage: pnpm send-welcome-email <email address>`;
